dashboard/requirements.txt
streamz
pyviz
colorcet
hvplot
panel
bokeh
azure-eventhub
azure-identity